atlas-ai/backend/requirements.txt
# Atlas AI — Python Backend Dependencies
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
pydantic==2.7.1
httpx==0.27.0

# Local AI
transformers==4.41.2
torch==2.3.0
accelerate==0.30.0
sentencepiece==0.2.0

# Twilio (optional — comm agent)
twilio==9.1.0

# Utilities
aiofiles==23.2.1
python-multipart==0.0.9